getter/pyproject.toml
[tool.poetry]
name = "agent"
version = "0.1.0"
description = ""
authors = ["Etienne <etienne.micheladj@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
redis = {extras = ["hiredis"], version = "^4.6.0"}
pyyaml = "^6.0.1"
metatrader5 = "^5.0.45"
pandas = "^2.1.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
